backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend API for Daedalus Dapp"
authors = [
    {name = "cassiel-st4vr0s",email = "csdmd.snf21@uea.edu.br"}
]
requires-python = ">=3.12,<4"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
    "web3 (>=7.12.0,<8.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
